Isolate client state and chain methods
