+++ crates/relayer-cosmos/src/methods/mod.rs
@@ -1,4 +1,6 @@
+pub mod chain;
 pub mod channel;
+pub mod client_state;
 pub mod connection;
 pub mod consensus_state;
 pub mod create_client;
